requirements.txt
Django>=2.0
django-debug-toolbar>=1.9.1
pytz
sqlparse
wheel
django-queryinspect>=1.0.0
uwsgi>=2.0.15
psycopg2>=2.7.1
django-easycart>=0.4.0
